# temp_openapi.json
{"openapi":"3.1.0","info":{"title":"Todo Chatbot API","description":"API for the AI-Powered Todo Chatbot with MCP integration","version":"1.0.0"},"paths":{"/api/v1/todos/":{"post":{"tags":["todos"],"summary":"Create Todo","description":"Create a new todo item.","operationId":"create_todo_api_v1_todos__post","parameters":[{"name":"description","in":"query","required":true,"schema":{"type":"string","title":"Description"}},{"name":"user_id","in":"query","required":true,"schema":{"type":"integer","title":"User Id"}},{"name":"due_date","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Due Date"}},{"name":"category","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Category"}},{"name":"priority","in":"query","required":false,"schema":{"$ref":"#/components/schemas/TaskPriority","default":"medium"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Response Create Todo Api V1 Todos  Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["todos"],"summary":"Get Todos","description":"Get all todos for a specific user with optional status filtering.","operationId":"get_todos_api_v1_todos__get","parameters":[{"name":"user_id","in":"query","required":true,"schema":{"type":"integer","title":"User Id"}},{"name":"status","in":"query","required":false,"schema":{"anyOf":[{"$ref":"#/components/schemas/TaskStatus"},{"type":"null"}],"title":"Status"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"type":"object","additionalProperties":true},"title":"Response Get Todos Api V1 Todos  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/todos/{task_id}":{"get":{"tags":["todos"],"summary":"Get Todo","description":"Get a specific todo item.","operationId":"get_todo_api_v1_todos__task_id__get","parameters":[{"name":"task_id","in":"path","required":true,"schema":{"type":"integer","title":"Task Id"}},{"name":"user_id","in":"query","required":true,"schema":{"type":"integer","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Response Get Todo Api V1 Todos  Task Id  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["todos"],"summary":"Update Todo","description":"Update a todo item's details.","operationId":"update_todo_api_v1_todos__task_id__put","parameters":[{"name":"task_id","in":"path","required":true,"schema":{"type":"integer","title":"Task Id"}},{"name":"user_id","in":"query","required":true,"schema":{"type":"integer","title":"User Id"}},{"name":"description","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"}},{"name":"due_date","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Due Date"}},{"name":"category","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Category"}},{"name":"priority","in":"query","required":false,"schema":{"anyOf":[{"$ref":"#/components/schemas/TaskPriority"},{"type":"null"}],"title":"Priority"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Response Update Todo Api V1 Todos  Task Id  Put"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["todos"],"summary":"Delete Todo","description":"Delete a todo item.","operationId":"delete_todo_api_v1_todos__task_id__delete","parameters":[{"name":"task_id","in":"path","required":true,"schema":{"type":"integer","title":"Task Id"}},{"name":"user_id","in":"query","required":true,"schema":{"type":"integer","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Response Delete Todo Api V1 Todos  Task Id  Delete"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/todos/{task_id}/status":{"put":{"tags":["todos"],"summary":"Update Todo Status","description":"Update the status of a todo item.","operationId":"update_todo_status_api_v1_todos__task_id__status_put","parameters":[{"name":"task_id","in":"path","required":true,"schema":{"type":"integer","title":"Task Id"}},{"name":"user_id","in":"query","required":true,"schema":{"type":"integer","title":"User Id"}},{"name":"status_update","in":"query","required":true,"schema":{"$ref":"#/components/schemas/TaskStatus"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Response Update Todo Status Api V1 Todos  Task Id  Status Put"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/todos/search":{"post":{"tags":["todos"],"summary":"Search Todos","description":"Search for todos by description.","operationId":"search_todos_api_v1_todos_search_post","parameters":[{"name":"user_id","in":"query","required":true,"schema":{"type":"integer","title":"User Id"}},{"name":"query","in":"query","required":true,"schema":{"type":"string","title":"Query"}},{"name":"status","in":"query","required":false,"schema":{"anyOf":[{"$ref":"#/components/schemas/TaskStatus"},{"type":"null"}],"title":"Status"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"type":"object","additionalProperties":true},"title":"Response Search Todos Api V1 Todos Search Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/mcp/create_todo":{"post":{"tags":["mcp"],"summary":"Create Todo Mcp","description":"MCP tool endpoint to create a new todo item.","operationId":"create_todo_mcp_api_v1_mcp_create_todo_post","parameters":[{"name":"user_id","in":"query","required":true,"schema":{"type":"integer","title":"User Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTodoRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/mcp/list_todos":{"post":{"tags":["mcp"],"summary":"List Todos Mcp","description":"MCP tool endpoint to retrieve a list of todo items with optional filtering.","operationId":"list_todos_mcp_api_v1_mcp_list_todos_post","parameters":[{"name":"user_id","in":"query","required":true,"schema":{"type":"integer","title":"User Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ListTodosRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/mcp/update_todo_status":{"post":{"tags":["mcp"],"summary":"Update Todo Status Mcp","description":"MCP tool endpoint to update the status of a specific todo item.","operationId":"update_todo_status_mcp_api_v1_mcp_update_todo_status_post","parameters":[{"name":"user_id","in":"query","required":true,"schema":{"type":"integer","title":"User Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateTodoStatusRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/mcp/delete_todo":{"post":{"tags":["mcp"],"summary":"Delete Todo Mcp","description":"MCP tool endpoint to remove a specific todo item.","operationId":"delete_todo_mcp_api_v1_mcp_delete_todo_post","parameters":[{"name":"user_id","in":"query","required":true,"schema":{"type":"integer","title":"User Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeleteTodoRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/mcp/search_todos":{"post":{"tags":["mcp"],"summary":"Search Todos Mcp","description":"MCP tool endpoint to search for todos by description or other criteria.","operationId":"search_todos_mcp_api_v1_mcp_search_todos_post","parameters":[{"name":"user_id","in":"query","required":true,"schema":{"type":"integer","title":"User Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SearchTodosRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/mcp/set_reminder":{"post":{"tags":["mcp"],"summary":"Set Reminder Mcp","description":"MCP tool endpoint to set a reminder for a specific task or event.","operationId":"set_reminder_mcp_api_v1_mcp_set_reminder_post","parameters":[{"name":"user_id","in":"query","required":true,"schema":{"type":"integer","title":"User Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SetReminderRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/chat/":{"post":{"tags":["chat"],"summary":"Chat","description":"Process a chat message and return an AI response.","operationId":"chat_api_v1_chat__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatMessageRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/chat/{conversation_id}":{"get":{"tags":["chat"],"summary":"Get Chat History","description":"Get chat history for a specific conversation.","operationId":"get_chat_history_api_v1_chat__conversation_id__get","parameters":[{"name":"conversation_id","in":"path","required":true,"schema":{"type":"integer","title":"Conversation Id"}},{"name":"user_id","in":"query","required":false,"schema":{"type":"integer","default":1,"title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/register":{"post":{"tags":["authentication"],"summary":"Register User","description":"Register a new user.","operationId":"register_user_auth_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/login":{"post":{"tags":["authentication"],"summary":"Login User","description":"Authenticate user and return access token.","operationId":"login_user_auth_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Read Root","description":"Root endpoint for health check.","operationId":"read_root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health":{"get":{"summary":"Health Check","description":"Health check endpoint.","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/user/{user_id}/conversations":{"get":{"summary":"Get User Conversations","description":"Get all conversations for a user.","operationId":"get_user_conversations_api_v1_user__user_id__conversations_get","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"integer","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"ChatMessageRequest":{"properties":{"message":{"type":"string","title":"Message"},"user_id":{"type":"integer","title":"User Id","default":1}},"type":"object","required":["message"],"title":"ChatMessageRequest"},"CreateTodoRequest":{"properties":{"description":{"type":"string","title":"Description"},"due_date":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Due Date"},"priority":{"type":"string","title":"Priority","default":"medium"}},"type":"object","required":["description"],"title":"CreateTodoRequest"},"DeleteTodoRequest":{"properties":{"task_id":{"type":"integer","title":"Task Id"}},"type":"object","required":["task_id"],"title":"DeleteTodoRequest"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ListTodosRequest":{"properties":{"status":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Status"},"priority":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Priority"},"limit":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Limit"}},"type":"object","title":"ListTodosRequest"},"LoginRequest":{"properties":{"email":{"type":"string","title":"Email"},"password":{"type":"string","title":"Password"}},"type":"object","required":["email","password"],"title":"LoginRequest"},"RegisterRequest":{"properties":{"name":{"type":"string","title":"Name"},"email":{"type":"string","title":"Email"},"password":{"type":"string","title":"Password"}},"type":"object","required":["name","email","password"],"title":"RegisterRequest"},"SearchTodosRequest":{"properties":{"query":{"type":"string","title":"Query"},"status":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Status"}},"type":"object","required":["query"],"title":"SearchTodosRequest"},"SetReminderRequest":{"properties":{"task_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Task Id"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"remind_at":{"type":"string","title":"Remind At"}},"type":"object","required":["remind_at"],"title":"SetReminderRequest"},"TaskPriority":{"type":"string","enum":["low","medium","high"],"title":"TaskPriority"},"TaskStatus":{"type":"string","enum":["pending","completed","in-progress"],"title":"TaskStatus"},"TokenResponse":{"properties":{"token":{"type":"string","title":"Token"},"user_id":{"type":"integer","title":"User Id"},"email":{"type":"string","title":"Email"}},"type":"object","required":["token","user_id","email"],"title":"TokenResponse"},"UpdateTodoStatusRequest":{"properties":{"task_id":{"type":"integer","title":"Task Id"},"status":{"type":"string","title":"Status"}},"type":"object","required":["task_id","status"],"title":"UpdateTodoStatusRequest"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}